re-add crazy glue and lube as guaranteed to toy box
in a previous change the toy box was changed to have random contents, but this makes glue and lube containers extremely difficult to obtain. simply re-adding them as a guarantee to the toybox crates allows clowns and scientists to still have consistency in sourcing materials while keeping the fun factor of the change.